Interactive Temperature Sensor Jan 18, 2015
Utilizes Arduino Uno
To start, activate the switch to begin checking for user. It will
continue if the temperature is above 23 degrees which is slightly warmer
than room temperature to prevent starting without a user.

A user will hold the temperature sensor for a given amount of time to
heat up and provide an accurate reading. The Arduino will read the input
values and convert them to a voltage reading then to temperature prior
to asking the user to release.

The temperature will be shown to the user on the liquid display lasting
a few seconds then it will take 30 seconds to restart. This will allow
the sensor to cool to room temperature to provide an accurate reading
for the next user.